Add files via upload
A simple converter which takes number of miles as an input and returns an output consisting of number of kilometres. 
An American friend of mine and I are car enthusiasts so its common for us to have to convert speed data from miles to kilometres so I decided to make one myself rather then using google. Used tkinter to make it look a bit more cooler.