Add files via upload
Simple Miles to Kilometres converter implemented with tkinter for coolness. 
An American friend of mine and I are car enthusiasts so it was common for us to convert values from miles and km and the other way around so I decided to make one myself rather than using google. Used tkinter for coolness only.